pyproject.toml
[project]
name = "converge"
version = "0.1.0"
description = "A distributed operating system for agents"
authors = [{name = "gsbm"}]
requires-python = ">=3.11"
dependencies = [
    "asyncio>=3.4.3",
    "cryptography>=41.0.0",
    "pydantic>=2.0.0",
    "msgpack>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "openai>=1.0",
    "anthropic>=0.18",
    "mistralai>=1.0",
    "ruff>=0.1.0",
    "pytest>=7.0.0",
    "coverage>=7.0.0",
    "pytest-asyncio>=0.23.0",
    "pyright>=1.1.408",
    "pre-commit",
    "pip-audit",
]
docs = [
    "sphinx>=7.0.0",
    "myst-parser>=2.0.0",
    "shibuya"
]
llm = [
    "openai>=1.0",
    "anthropic>=0.18",
    "mistralai>=1.0",
]
websocket = [
    "websockets>=14.0",
]
cli = [
    "pyyaml>=6.0",
]

[project.scripts]
converge = "converge.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
ignore = []
extend-select = [
  "F",   # Pyflakes rules
  "W",   # PyCodeStyle warnings
  "I",   # Sort imports properly
  "UP",  # Upgrade syntax for newer Python
  "C4",  # Comprehension optimizations
  "FA",  # from __future__ import annotations
  "ISC", # Implicit string concatenation
  "ICN", # Import conventions
  "RET", # Return best practices
  "SIM", # Simplifications
  "TID", # Import tidiness
  "PTH", # pathlib preferred
  "TD",  # TODO comments
  "A",   # Shadowing builtins
  "COM", # Trailing commas
  "DTZ", # Timezone-aware datetime
  "FBT", # Boolean trap
  "N",   # Naming conventions
]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]
asyncio_mode = "auto"

[tool.coverage.run]
branch = true
source = ["converge"]
omit = [
    "converge/__init__.py",
    "*/__init__.py",
    "converge/__main__.py",
    "converge/cli.py",
    "converge/network/transport/tcp.py",
    "converge/network/transport/websocket.py",
]

[tool.coverage.report]
show_missing = true
fail_under = 98

[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "basic"
reportMissingTypeStubs = false
reportMissingImports = false  # Ignore missing imports for third-party packages without type stubs
reportUnusedImport = true
reportUnusedVariable = true
# Ignore third-party library type issues and known dynamic attribute access patterns
reportAttributeAccessIssue = false
reportCallIssue = false
reportReturnType = true
reportArgumentType = true
exclude = [
    "**/__pycache__",
    "**/.venv",
    "**/docs",
    "**/tests",
]
